Rework test directory, add test which uses move std lib
